deploy_requirements.txt
google-cloud-run==0.10.5
google-cloud-storage==2.14.0
google-cloud-resource-manager==1.12.0
google-api-python-client==2.108.0
pyyaml==6.0.1
python-dotenv==1.0.0

